Move LowerBound and UpperBound structs in uv-pep440
